% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_network.r
\name{R2_network}
\alias{R2_network}
\title{R2 network connectedness}
\usage{
R2_network(
  data,
  method = "genizi",
  directed = TRUE,
  amat = FALSE,
  prev_amat = NULL
)
}
\arguments{
\item{data}{data.frame or matrix with column names}

\item{method}{character, default "genizi"}

\item{directed}{logical}

\item{amat}{FALSE or adjacency matrix (child x parent)}

\item{prev_amat}{optional (currently unused)}
}
\value{
A list with network tables and summary measures
}
\description{
R2 network connectedness
}
