% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_dy_network.r
\name{rolling_dy_network}
\alias{rolling_dy_network}
\title{Rolling Diebold-Yilmaz connectedness (TCI)}
\usage{
rolling_dy_network(data, block_size, n.ahead)
}
\arguments{
\item{data}{data.frame or matrix}

\item{block_size}{Integer rolling window size}

\item{n.ahead}{Integer horizon (forecast steps used in FEVD)}
}
\value{
A list containing a rolling TCI vector.
}
\description{
Rolling Diebold-Yilmaz connectedness (TCI)
}
