% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_irf_network.r
\name{rolling_irf_network}
\alias{rolling_irf_network}
\title{Rolling IRF network}
\usage{
rolling_irf_network(data, block_size, n.ahead, cumsum = TRUE, amat = FALSE)
}
\arguments{
\item{data}{data.frame or matrix}

\item{block_size}{Integer rolling window size}

\item{n.ahead}{Integer horizon}

\item{cumsum}{Logical; if TRUE uses cumulative IRFs}

\item{amat}{FALSE or adjacency matrix passed to irf_network}
}
\value{
A list containing a rolling TCI vector and rolling TO/FROM data frames.
}
\description{
Rolling IRF network
}
