% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_R2_network.r
\name{rolling_R2_network}
\alias{rolling_R2_network}
\title{Rolling window R2 network analysis}
\usage{
rolling_R2_network(
  data,
  window_size,
  step = 1,
  bootstrap = 1,
  amat_list = FALSE,
  ...
)
}
\arguments{
\item{data}{data.frame or matrix}

\item{window_size}{integer window length}

\item{step}{integer step size}

\item{bootstrap}{number of bootstrap replications}

\item{amat_list}{FALSE or list of adjacency matrices}

\item{...}{passed to R2_network}
}
\value{
A list of rolling network measures and tables
}
\description{
Rolling window R2 network analysis
}
