% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_R2_network.r
\name{theoretical_R2_network}
\alias{theoretical_R2_network}
\title{Theoretical values of the R2 network}
\usage{
theoretical_R2_network(
  A0,
  var_names = paste0("X", 1:nrow(A0)),
  directed = TRUE
)
}
\arguments{
\item{A0}{Structural matrix (p x p)}

\item{var_names}{Variable names}

\item{directed}{logical}
}
\value{
A list with theoretical network tables and measures
}
\description{
Theoretical values of the R2 network
}
